requirements.txt
confluent-kafka
reduct-py